CacheStorage: Support large file on Cache Storage API
This is a last cl required for supporting 2GB+ file.

Bug: 391398191
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6705024
Commit-Queue: Eriko Kurimoto <elkurin@chromium.org>
Auto-Submit: Eriko Kurimoto <elkurin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1500000}

+++ content/browser/cache_storage/cache_storage_cache_entry_handler.h
@@ -87,7 +87,7 @@ class CacheStorageCacheEntryHandler {
              int bytes_to_read,
              base::OnceCallback<void(int)> callback);
 
-    int GetSize(CacheStorageCache::EntryIndex disk_cache_index) const;
+    int64_t GetSize(CacheStorageCache::EntryIndex disk_cache_index) const;
 
     void Invalidate();
